+++ src/client/model/Entity.ts
