+++ src/provider/model/Result.ts
