+++ src/provider/view/components/Notification/Notification.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
 import {NotificationTime} from '../NotificationTime/NotificationTime';
-import {NotificationCenterAPI} from '../../../models/NotificationCenterAPI';
+import {NotificationCenterAPI} from '../../../model/NotificationCenterAPI';
 import {Button} from '../Button/Button';
-import {NotificationTypes} from '../../../../client/models/Notification';
-import {INotification} from '../../../models/INotification';
+import {NotificationTypes} from '../../../../client/model/Notification';
+import {INotification} from '../../../model/INotification';
 declare var window: Window & {openfin: {notifications: NotificationCenterAPI}};
